requirements/base.txt
argon2-cffi==18.1.0
awesome-slugify==1.6.5
Pillow==5.1.0
requests==2.19.1

# Django
Django==2.0.6
django-environ==0.4.4
django-model-utils==3.1.2

# Django Rest Framework
djangorestframework==3.8.2
coreapi==2.3.3
